BioFormatsFormat: compute reader classes only once
This change reworks how ImageReaders are created. The ClassList for the
reader classes is computed once when the format is first created;
afterward, that same ClassList is always used to create new
ImageReaders.

This change also avoids calling close() on the ImageReader when such a
call is unnecessary; this substantially improves code conciseness.