Add SLF4J binding dependency
Since loci:scifio:4.4.8 uses SLF4J 1.5.10, it requires an SLF4J binding
at runtime or else it will throw NoClassDefFoundError. This is fixed in
SLF4J 1.6; hopefully the 4.4.9 release of Bio-Formats will use a newer
SLF4J and then we won't need to worry about it anymore. But in the
meantime, we add a test-scope dependency on slf4j-log4j12 so that the
ReadFile program in src/test runs as expected. Any downstream consumers
of scifio-bf-compat will, for the time being, also need to add an SLF4J
binding (not necessarily log4j though) as a dependency.