Metadata implements HasColorTable
The Bio-Formats Format's metadata should implement HasColorTable. We
were setting the color table of the plane during openPlane calls via the
reader, but it should have been attached to the Metadata so other
aspects of the framework will find it where expected.

This also allows us to cache it in a logic way.