Properly transfer channel axes
Fundamentally we have three Channel scenarios that need to be
represented in SCIFIO:
* Interleaved RGB channels
* Non-interleaved RGB channels
* Non-planar channels

In attempting to preserve sub-channel labels, we disregarded the
RGBChannelCount and EffectiveSizeC lengths, which lead to incorrect
SCIFIO Metadata population.

Updated so that RGBChannelCount maps to a planar Axes.CHANNEL axis
(interleaved or not depending on reader state).

EffectiveSizeC maps to a non-planar channel axis.

If the reader is indexed, there will be a guaranteed planar Axes.CHANNEL
axis if there effectiveSizeC == 1... thus we assume that the indexed
axes are RGB, unless otherwise specified.

Note that with this change we do not consider sub-channel labels.