Improve color table support
If a reader does not have a ColorTable, but we do have a Color object in
the Bio-Formats metadata, we can create a black : color ramp using that
Color.

NB: Color is an 8-bit color and thus must be chunked to get a 16-bit
color table.

NB: the color table is currently not being cached.