Do not attempt to load non-File instances via Bio-Formats
There is apparently code in Bio-Formats that can open random access
streams that are not backed by local files, but we do not use that
code path anyway (at least if this developer understood Curtis
Rueden's generous education correctly).

So let's not have tons of attempts by each Bio-Formats reader to
read something that we know it cannot read, at least the way we
asked them to read it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>