Move io.scif.common package
To facilitate OSGi builds, which do not allow a package to appear in
more than one bundle, io.scif.common is moved to io.scif.bf.wrapper.
This also establishes a common base package of io.scif.bf for this
component.

The io.scif.common.RandomAccessInputStreamWrapper is deprecated and
delegates to the io.scif.bf.wrapper implementation.