Use cached ImageReader when possible
Constructing a new ImageReader imposes a non-trivial performance hit, as
it calls ReaderClass.newInstance for each Bio-Formats reader.

For the purposes of generating suffixes or checking for format
compatibility, this is completely unnecessary if the class list hasn't
changed since the last time it was queried.

Thus we now check the hash of the ClassList contents and only
re-generate ImageReaders if a change is detected.