Cache the metadata store
There's no reason to expect the Reader will be open after parsing the
Metadata. This is a problem, for example, with the OMETIFFReader which
- if closed - will throw an IllegalStateException when retrieving the
metadata store.

To mitigate this risk we cache the metadata store, and clear the cache
when the reader changes.

Required for https://github.com/scijava/pom-scijava/issues/123